requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
tqdm>=4.66.0
PyYAML>=6.0.1
opencv-python>=4.8.0
Pillow>=10.0.0
transformers>=4.41.0    
ftfy>=6.1.1             
regex>=2024.5.15        
scikit-image>=0.22.0    